# Adding two level hashing in metrics hashmap 