## Changes

This PR is opened just for the discussion, of one of the possible changes to improve the performance of concurrent recording/updation of measurements in the hot-path. In today's (Feb 20th, 2024) community meeting, we discussed multiple approaches 
- Using [`dashmap`](https://crates.io/crates/dashmap) - sharded hashmap. 
- Using [`flurry`](https://crates.io/crates/flurry) - uses fine-grained atomic locks in hashmap
- Using [`sharde-slab`](https://crates.io/crates/sharded-slab) - building hashmap using concurrent data-structures provided by this crate.
- Using `thread-local` hashmaps in the hot-path to record measurements and then merge them during collection. The collection would be a performance-heavy process. And the SDK doesn't have control over the life cycle of threads created by the application, so exiting a thread by the application will also remove all it's aggregated data.
- Cijo spoke about the lock-free approach for hot-path as used in otel-dotnet, where the first level of hashmap only stores the index to the aggregated data. 

The approach in this PR is modifying the `ValueMap` store values in two-level hashing to minimize lock contention. It's basically a more simpler form of sharding as provided by `dashmap`.

Existing:

```rust
struct ValueMap<T: Number<T>> {
    values: Mutex<HashMap<AttributeSet, T>>, 
    has_no_value_attribute_value: AtomicBool,
    no_attribute_value: T::AtomicTracker,
}
```

PR:
```
struct ValueMap<T: Number<T>> {
    buckets: Arc<HashMap<u8, Mutex<HashMap<AttributeSet, T>>>>, // ? or RWLock
    has_no_value_attribute_value: AtomicBool,
    no_attribute_value: T::AtomicTracker,
    total_count: AtomicUsize,
}
```

First level of hashing is to distribute the values across the fixed set of 256 buckets.  Each bucket is guarded by its own mutex and contains a second-level hash map for storing AttributeSet to aggregation mapping. 

Please note
 - This approach won't see any improvement in single-threaded application, as we still lock. Infact, the complexity added by sharding may slightly impact the performance. I didn't see any perf improvement/degradation in benchmark tests as they are single threaded.
 - If all the attributes are hashed to the same bucket ( rare scenario), there won't be any perf improvement as all the threads will contest for the same lock.
 - This approach may impact [`locality of reference`](https://en.wikipedia.org/wiki/Locality_of_reference) during collection cycle, as the OS can allocate these buckets in different segments of memory, and so mayn't be cached together.
 
All above are the limitation of any sharded hashmap (including `dashmap`).

This is the result of our metrics stress test in my machine. However, I don't think results are really consistent across machines, so good if someone can test it too:

| Threads  | Main Throughput (millions/sec)   | PR  Throughput (millions/sec)  | PR Throughput (millions/sec) with hashbrown |
| ---|---|---|---|
| 3  | 6.6  | 6.6   | 6.8
| 5  | 10.1  |12.5   | 13.2
| 8 | 5.6  |  18.9 | 20.0
| 10 |  4.5 | 22.0  |  24.0


Based on the above results- with PR, the perf seems to be increasing significantly with the number of threads. In the main branch, the performance increases with threads till a threshold and then starts degrading.

Benchmark result. This PR won't improve performance there, as these are single-threaded tests. There is slight retrogress for two-level indirection, which gets compensated (in fact improved) with hashbrown/ahash:


| Test | time (Main branch) | time (PR branch) | time (PR with hashbrown) |
| ---|---|---|---|
| AddNoAttrs | 30.994 ns  | 31.286 ns  |31.544 ns
|AddNoAttrsDelta  | 31.009 ns  |31.009 ns   | 31.455 ns
| AddOneAttr | 169.76 ns |  184.00 ns| 154.39 ns
| AddOneAttrDelta |172.54 ns| 185.37 ns  |  157.08 ns
| AddThreeAttr |396.93 ns |393.36 ns | 368.69 ns| 
| AddThreeAttrDelta |396.41 ns | 393.68 ns|368.60 ns |
| AddFiveAttr |706.21 ns |718.80 ns |672.07 ns | 


Hashbrown/ahash seems widely adopted crates in terms of number of downloads, and also secure enough for external ddos attacks - https://github.com/tkaitchuck/aHash/wiki/How-aHash-is-resists-DOS-attacks


## Merge requirement checklist

* [ ] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [ ] Unit tests added/updated (if applicable)
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [ ] Changes in public API reviewed (if applicable)