@hdost 2024-02-21
@lalitb love the performance gains! Were you able to test this impl against some of the single level hashmap alternative implementations?

> Lets sit on this for **1- 2 days to see if any concerns**, before proceeding to do similar for other instrument types. Also, lets do this after the [pending release](https://github.com/open-telemetry/opentelemetry-rust/pull/1539).

@cijothomas If it's truly for discussion a week seems like a more reasonable timeframe.