@lalitb 2024-02-22
> Were you able to test this impl against some of the single level hashmap alternative implementations?

@hdost Do you have any alternative implemetations in mind, can test that.

I tested `dashmap` (which is also a sharded hashmap) but the performance was worse than our existing implementation - https://github.com/open-telemetry/opentelemetry-rust/compare/main...lalitb:metrics-dashmap?expand=1


`main branch (using hashmap)`:
Number of threads: 8
Throughput: 6,640,600 iterations/sec
Throughput: 6,203,800 iterations/sec

`using dashmap`:
Number of threads: 8
Throughput: 1,096,800 iterations/sec
Throughput: 1,184,600 iterations/sec

In general, as tested separately, the dashmap was much faster for concurrent read-only operations, but our scenario is concurrent updates. Also,  we were using dashmap in the old metrics implementation, but was removed in the new PR  (didn't find the reason in history). In general, reluctant to take dependency on any external crate unless it is widely used :)