@hdost 2024-02-22
From what it seems there's no public interface effect with this so let's take the win for now and see if we can't make further improvements later 👍