@lalitb 2024-02-28
Also added benchmark results, they are single-threaded, so the changes won't improve those., though we see the improvement with the use of hashbrown/ahash.