@lalitb 2024-07-02
There have been substantial perf improvements in metrics instrumentation with #1833.  Will revisit this PR after logs beta release. Closing this for now to keep the PR list minimal.